Core/Crypto: Rename our Argon2 to avoid possible conflict with library file on some configurations
(cherry picked from commit 4f6b49cccf87638864a636989a41f9ca7eb7a89c)
