+++ src/common/Cryptography/Argon2Hash.cpp
@@ -15,7 +15,7 @@
  * with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "Argon2.h"
+#include "Argon2Hash.h"
 #include <argon2/argon2.h>
 
 /*static*/ Optional<std::string> Trinity::Crypto::Argon2::Hash(std::string const& password, BigNumber const& salt, uint32 nIterations, uint32 kibMemoryCost)
